AddressData: extracted from SegwitAddressTest
Put it in base.internal so it can also be used by tests
in the core module.
